third_party/WebKit/LayoutTests/fast/html/imports/import-style-tree-order-expected.txt

PASS window.getComputedStyle(styleShouldBeHiddenByImport).color is "rgb(0, 128, 0)"
PASS window.getComputedStyle(linkShouldBeHiddenByImport).color is "rgb(0, 128, 0)"
PASS window.getComputedStyle(importShouldBeHiddenByImport).color is "rgb(0, 128, 0)"
PASS window.getComputedStyle(subimportShouldBeHiddenByImport).color is "rgb(0, 128, 0)"
PASS window.getComputedStyle(sharedImportShouldBeIgnored).color is "rgb(0, 128, 0)"
PASS window.getComputedStyle(styleShouldBeHiddenByDynamicallyAddedFollower).color is "rgb(0, 128, 0)"
PASS window.getComputedStyle(styleShouldNotBeHiddenByDynamicallyAddedLeader).color is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
